eos-update-flatpak-repos: avoid rewriting exports un-necessarily
047f46b1 introduced the idea of unconditionally rewriting the metadata in the
case that the app being migrated is an extra data app, previously only done
when the app ID was changing. Unfortunately this exposed a bug in the rewriting
of exports which causes the file handling to fail when the new and old ID are
the same. Avoid entering this code path if the app IDs are the same.

https://phabricator.endlessm.com/T26588