image-boot: Create partition for persistent data storage
In order to add persistent data storage capabilities to our
live+installer setup (so that you don't lose your data on reboot),
add another step in the boot process that creates the partition.

Similar to the repartitioning case for first-boot of installed systems,
this must be done in the initramfs, before any filesystems are mounted,
so that the kernel is able to update it's view of the partition table
without requiring a reboot.

Even though there are some similarities to endless-repartition.sh, the
actual details are quite different so I found it more appropriate to
implement as a separate script.

The actual filesystem creation is deferred to the boot from the "real
root" and will be handled separately.

https://phabricator.endlessm.com/T27158