image-boot: alternative method for seeding machine-id in overlay
The approach attempted in commit 4d348fb9b56ff34cf150d68a6012b6ce5778d51a
didn't work because the parent mount at /run is a shared mount, and
mount --move cannot be used there.

There isn't an existing clear place in the hierarchy where we can work
with private mounts instead, so just change the approach, instead
populating the overlayfs upperdir in advance.

This will need tweaking when we implement support for persistent overlayfs
storage (and that's also true for other code in this file).

https://phabricator.endlessm.com/T27158