image-boot: improve robustness of iso9660 partition detection
During live image boot, eos-image-boot-generator attempts to set up
the systemd units so that they wait for the relevant storage device
to become available, then eos-image-boot-setup will be executed to work
on that device.

However, in the case of booting an ISO flashed to USB, there is some
ambiguity here: both the USB disk (e.g. sda) and the iso9660 partition
(e.g. sda1) are assigned the ISO's UUID, so the target of the
/dev/disk/by-uuid/ link varies during the boot process.
Commit de8800c0bcd055a previously identified and worked around this.

The workaround is not ideal though, and this situation gets harder again
as we look to add code that creates a persistent data partition. Right
after creation, the unformatted persistent data partition will also inherit
the ISO UUID, so the by-uuid link changes for a second time.

Switch to a more robust scheme, taking advantage of blkid which already
has the logic we want: it can search by UUID for partitions that actually
have a filesystem. The USB disk (sda) does not, and once formatted, the
persistent data partition will have differing a UUID originating from the
filesystem. Using blkid looks like it will reliably pick out the iso9660
partition that we want to mount.

Also remove support for endless.image.device=PARTUUID= which is not
covered by this new scheme and is not used by existing boot paths.