tests: fix running from parent directory
tests/run-tests is useful for running all the tests, however the only way
I can find to run a specific tests is running:

 python3 -m unittest tests.test_foo.bar

from the main eos-boot-helper directory.

However, the reclaim-swap tests fail here, because they look for the
data in the wrong place.

Make the test more robust in finding the accompanying data.