image-boot-setup: mark main partition read-only after repartitioning
The block device re-probing caused by eos-live-storage-setup has the
side effect of losing the read-only bit that we had set beforehand.

Once the graphical session has loaded, that partition can no longer
be mounted (fails with -EBUSY?) unless you mount with "-o ro".
The installer runs into this error and hence cannot install Endless.

Reorder so that we continue booting with this partition marked
as read-only.

https://phabricator.endlessm.com/T27158