image-boot-setup: avoid strange blkid lookup behaviour
For some reason, when blkid is run with --uuid it enters a weird
"eval" mode that behaves differently from other blkid commands.
The eval mode does not have the behaviour that I was trying to
rely on here, so tweak the parameters to make blkid really behave
as I describe.

Filed upstream bug report:
https://github.com/karelzak/util-linux/issues/866

This fixes an issue where the system would boot and the by-uuid
symlink would sometimes point to sdb3 (unformatted persistent storage
partition). The previous blkid command's unwanted behaviour would then
direct us to that as the main data partition, causing boot to fail.

https://phabricator.endlessm.com/T27158