live-boot: format and use persistent storage partition
On ISO-USB boots, an extra partition will now be created during early
boot.

Once we reach the point of overlayfs setup, attempt to locate this
partition.

If it is unformatted (still has the marker), format it as ext4.

Then mount it and use it as persistent backing storage for the
overlayfses.

https://phabricator.endlessm.com/T27771
(cherry picked from commit 874fef628748521793bde6528eec0a7a878bdec4)