live-boot: create flatpak overlayfs on subsequent boots too
Now that /run/eos-live has persistent backing storage for
ISO USB boots, we need to go ahead and create this overlayfs
even when the upperdir already exists.

Otherwise the flatpak setup is only done on first boot, but not on
any subsequent boots of this media, and you can't launch your flatpak
apps as a result.

https://phabricator.endlessm.com/T27771