hack: Add new service to remove the global override
Old hack computer has a global flatpak override file. In the new hack
experience we're going from using a custom product EOS to use the main
EOS branch so, we need to remove this global override and we'll use
local override for each application.

This patch adds a service that will remove the global override, only for
old hack-computer users, it checks for the HackComponents flatpak that
should only be installed in old hack computers, and if that exists it'll
remove the global override.

This service is based on the eos-update-flatpak-repos to be run only on
update.

https://phabricator.endlessm.com/T27984