eos-update-flatpak-repos: Migrate GNOME 3.20 and fd.o 1.4
The now-retired "gnome" flatpak remote served org.gnome.Platform//3.20
and org.freedesktop.Platform//1.4, so there's some chance these exist on
some user's system, and are preventing that user from updating due to
https://phabricator.endlessm.com/T25809. So migrate them both to
Flathub. Flathub does not in fact serve those runtimes any more (or
ever?) but as long as they're not deployed from a remote that doesn't
exist, the update to 3.3.20 (yet to be released) and beyond should
succeed.

https://phabricator.endlessm.com/T25809