eos-update-flatpak-repos: Fix remote URL snafu
During the 3.7.6 release I made a poorly thought through decision to
override the image builder URLs so that they pulled directly from our
server instead of our CDN that kept failing. I thought that the URLs
would only be used while pulling the data and not end up as the final
URLs. That was not the case.

Add a function to walk through all the OSTree and Flatpak remotes and
swap our internal ostree.endlessm-sf.com domain for the CDN
ostree.endlessm.com domains.

https://phabricator.endlessm.com/T29535