eos-live-boot-overlayfs-setup: combo usb stick persistent storage
This lets us use a file written on the exfat partition of a "combo" usb
stick as persistent storage.

The installer will fill the exfat partition of these devices with
a file named "persistent.img" which contains the same marker we
use in the unformatted ext4 partitions of other live usb sticks.

https://phabricator.endlessm.com/T14891