dracut: Add the device mapper module to our config
We use the device mapper in our dual boot configurations to map the file
on the user's windows partition as a block device that hosts endless.

Recently something broke in these configurations, and now systemd tries
to unmount our /var partition at boot. It appears that this may have
been a problem waiting to manifest for some time, because under 3.7.8
we see evidence of systemd wanting to unmount /var but bailing on the
process due to other conflicts.

The problem is that on the transition from initrd to real root we
have to clean the udev db to remove anything that was created with
the initrd's set of udev rules - which may differ from the booted
system's udev rules.

Once the db is cleared, udevadm trigger is used to introduce all
the devices to the booted system's ruleset.

However, device mapper entries have a DM_UDEV_PRIMARY_SOURCE flag to
indicate that they're not synthetic, and this flag is lost when the
udev db is recreated from a trigger. At this point the now seen as
synthetic events trigger some other udev rules which eventually
lead to SYSTEMD_READY=0 being added for the device.

That in turn causes systemd to see the device as not ready, which
moves it from (mounted | seen_by_udev) to (mounted). It interprets
this transition as the device having gone away, so unmounts any
filesystems it knows are on it.

Somehow that I can't remember or didn't figure out this doesn't
unmount our root partition, but it gets our /var mount which is
mounted by a generator provided by ostree.

In order to prevent this from happening we need to add the db_persist
option to all the device mapper block devices in udev. This causes
the udevadm info --cleanup-db call during startup to leave all the
device mapper block devices as they are. However, that rule should only
be present during initrd, so doing another udevadm info --cleanup-db
as root after boot can appropriately mess up the running system.

Instead of making these rules ourselves, we can pull them in by
making our dracut image-boot module depend on the "dm" module which,
among other things, adds the appropriate db_persist option for
device mapper devices.

We also get to drop some rules we explicitly pulled in that are
now pulled in by "dm"

https://phabricator.endlessm.com/T30165