image-boot-dm-setup: Add a hole for winstaller persistent storage
We've added a new file to the "combo" USB sticks created by the
windows install tool. This file is used for persistent storage, and
is mapped by the device mapper.

That means we need to add it to the hole image-boot-dm-setup makes
so we can mount the rest of the exFAT partition at install time
when we need it.

https://phabricator.endlessm.com/T30318