Only add Encyclopedia content symlink if appropriate
Previously, the tmpfiles.d snippet that makes a directory within
/run/media/eoslive appear as the com.endlessm.encyclopedia.en.Content
extension was run unconditionally. This means that on all systems,
/var/lib/flatpak/extension/com.endlessm.encyclopedia.en.Content/x86_64/eos3
was a symlink to /run/media/eoslive/.kiwix/flatpak.

On non-live systems, or on live systems which don't have this unusual
configuration, this causes the encyclopedia to fail to start:

    wjt@camille:~/src/endlessm/eos-boot-helper$ com.endlessm.encyclopedia.en
    bwrap: Can't find source path /var/lib/flatpak/extension/com.endlessm.encyclopedia.en.Content/x86_64/eos3: No such file or directory
    error: ldconfig failed, exit status 256

Move the tmpfiles.d snippet to a different directory, and only invoke it
if it is reasonable to do so: this is a live system, and the target
directory exists.

https://phabricator.endlessm.com/T30599