eos-update-flatpak-repos: Fix a missing ref edge case
A user on the forum hit this error:

INFO:root:Found gnome:runtime/org.gnome.Platform/x86_64/3.28 to migrate to flathub:runtime/org.gnome.Platform/x86_64/3.28
ERROR:root:Failure applying migration to org.gnome.Platform
Traceback (most recent call last):
  File "/sysroot/home/irena/eos-update-flatpak-repos", line 1698, in _migrate_installed_flatpaks
    migrate_extra = is_extra_data(repo, ostree_refs[old_ostree_ref])
KeyError: 'gnome:runtime/org.gnome.Platform/x86_64/3.28'

So, make the code a bit more robust so that it doesn't error out when
the old ref doesn't exist and we can still migrate the deploy file.
We had already encoded the assumption that the old ref might not exist
with the check "if old_ostree_ref in ostree_refs and not new_ref:", so
this just makes it so we don't error out in that case.

Reference:
https://community.endlessos.com/t/issues-with-updates-i-am-not-able-to-update-anything/14838

https://phabricator.endlessm.com/T31079