eos-update-flatpak-repos: Use the correct kind for app extensions
Without this change, having matched
runtime/com.hack_computer.Clubhouse.Locale/x86_64/eos3, the script would
try and fail to uninstall
app/com.hack_computer.Clubhouse.Locale/x86_64/eos3.

'kind' in this block previously referred to the kind of the app in
'FLATPAKS_TO_MIGRATE' in the outer 'for' loop, not the kind of the
matched Flatpak in the inner 'for' loop.

Reassign 'kind' in the inner loop, just as we do for 'name'.

https://phabricator.endlessm.com/T31185