tmpfiles: Ensure /etc/chromium-browser exists before attempting to create a symlink to it
This directory used to be created by the chromium-browser system package (which is
no longer present) and is used when setting up the org.chromium.Chromium.Policy.system-policies
flatpak extension. The lack of this directory breaks the extension and in turn
the chromium flatpak fails to startup. Let's fix it by ensuring the dir exists
when setting up the extension.

https://phabricator.endlessm.com/T30997