eos-update-flatpak-repos: Don't try to migrate partially-installed runtime
com.hack_computer.Clubhouse.Locale cannot currently be migrated from eos-apps
to flathub together with the main app. The current implementation fails
because the runtime is partially installed – i.e. only a subset of the
directories are present, corresponding to the locales on the system.

The .Locale runtime is optional, and does not contain any (meaningful)
translations, so the app works fine without it.

Work around the limitation by removing the runtime before migrating the app.
When Clubhouse is next updated from Flathub, either automatically or manually,
the .Locale extension will be re-installed.

https://phabricator.endlessm.com/T31185