eos-image-boot-dm-setup: Dump persistent.img extents separately
The error checking logic here relies on only one thing being done in
each branch of the case statement, since it looks at the return value
from the last commant ($?).

This commit moves the dumpexfat of persistent.img out of the case
statement, so the error checking works as designed.