chromium: Don't crash if .local/share/applications doesn't exist
os.listdir() raises FileNotFoundError if the given directory doesn't
exist. Previously, eos-migrate-chromium-profile would crash in this
case, such as for every new user profile:

    Traceback (most recent call last):
      File "/usr/lib/eos-boot-helper/eos-migrate-chromium-profile", line 150, in <module>
        main()
      File "/usr/lib/eos-boot-helper/eos-migrate-chromium-profile", line 132, in main
        update_desktop_shortcuts()
      File "/usr/lib/eos-boot-helper/eos-migrate-chromium-profile", line 110, in update_desktop_shortcuts
        for filename in os.listdir(DESKTOP_SHORTCUTS_DIR):
    FileNotFoundError: [Errno 2] No such file or directory: '/sysroot/home/erikos/.local/share/applications'

In the case where the user had an old Chromium profile at
~/.config/chromium but no ~/.local/share/applications, this would mean
that their Chromium profile would not be migrated to the
Flatpak-compliant location.

An alternative would be to use:

    for dirname, dirnames, filenames in os.walk(path):
        for filename in filenames:
            if filename.endswith(".desktop):
                update_desktop_shortcut(os.path.join(dirname, filename))

since os.walk() gracefully yields no entries if the path doesn't exist.
However this would change the meaning of this function to recursively
walk the directory. That's probably OK but I haven't thought through the
implications.

https://phabricator.endlessm.com/T31446