Add plumbing to add openh264 flatpak runtime to library search path when available
This change adds a tmpfiles.d config to link the openh264 flatpak
runtime at /run/lib/openh264, a ld.so.conf.d config to make this dir
available in the library search path and systemd units to re-run
ldconfig when the runtime is installed/uninstalled in order to make it
to the ldconfig cache.

We use a watcher service in order to re-run ldconfig (restart the unit
responsible to run it) as path units can only start other units, not
restart them.

https://phabricator.endlessm.com/T23316