Add tool to split flatpak repo out of system ostree repo
This tool splits the flatpak repo resources out of /ostree/repo and sets
up /var/lib/flatpak/repo up as a normal bare-user-only flatpak repo.
This is done by duplicating the repo and then pruning out the unwanted
components of the new flatpak repo and the original ostree repo.

For testing a directory closely resembling a real system is used, but
without absolute symlinks. It would be nice to do that, but it would
require setting up a bunch of mounts and chrooting.

https://phabricator.endlessm.com/T31877