eos-update-flatpak-repos: Migrate Solitaire & Freecell
These apps are actually the same codebase, and on Flathub there is only
one, Aisleriot.

This migration is imperfect. If both Freecell and Solitaire are
installed, only the first one will be listed in X-Flatpak-RenamedFrom in
the deployed copy of the migrated app. When the app is updated, both
will be added, since they are correctly both added to previous-ids, and
those get re-added to the .desktop file on each deploy.

I don't think this is too serious a problem. Necessarily we are
replacing two launchers with one. If a user has the "wrong" one pinned
to their taskbar and the icon disappears… oh well!  Solitaire is listed
first because it is somewhat more widely used than Freecell according to
our metrics.

https://phabricator.endlessm.com/T23936