eos-update-flatpak-repos: Unconditionally rewrite metadata
Previously, we would only update the metadata file, and assign it to
xa.metadata on the commit object, if the app ID had changed or if the
app was extra-data.

However, there are old apps which are missing xa.metadata on their
commit object. Trying to deploy such a commit is now a fatal error¹:

    INFO:root:Deploying with new ref: flathub:app/com.endlessm.photos/x86_64/stable
    ERROR:root:Failure applying migration to com.endlessm.photos
    Traceback (most recent call last):
      File "./eos-update-flatpak-repos", line 1784, in _migrate_installed_flatpaks
        arch, new_branch)
    gi.repository.GLib.GError: g-io-error-quark: Commit metadata for app/com.endlessm.photos/x86_64/stable not matching expected metadata (14)

To handle this case, update the metadata file and assign it to
xa.metadata for all apps. No changes are needed to rewrite_metadata(),
since it already handles the case where the app ID has not changed and
the app is not extra-data, and essentially becomes a no-op.

¹ flatpak-dir.c actually has a special-case for OCI remotes, which
  makes the absence of xa.metadata non-fatal. This is for backwards
  compatibility with old builds of apps where this was missing. So,
  if it is not actually needed, why ever make it fatal? Anyway…

https://phabricator.endlessm.com/T23936