split-flatpak-repo: Only hardlink objects and deltas
When duplicating the ostree repo we want to use hardlinks for speed and
to keep disk usage constant. However, hardlinking repo metadata files
might lead to errors or corruption if they're later modified without
replacing. Although ostree is generally careful about replacing files
like `config` when they're modified, there are no guarantees.

Similarly, some files are expected to be unique to each repo and would
not behave correctly when hardlinked. In particular, hardlinking the
repo `.lock` file would mean that locking one repo would lock both since
they're the same file at the inode level.

Instead, only hardlink the `objects` and `deltas` directories of the
ostree repo. This is where the vast majority of the data is and also
where ostree is guaranteed to operate in a replace-only manner. It's not
expected that any repos have static deltas, but copying instead of
linking them could significantly affect disk usage.

https://phabricator.endlessm.com/T32090