systemd preset: disable unwanted VPN services
The corresponding packages are brought in as dependencies of
network-manager-l2tp. However, I have checked the network-manager-l2tp
source code, and it looks like that plugin launches the required daemons
directly when needed.

There should be no reason to have these two services running all of the
time for all users.