overlayfs-setup: Remove several unneeded overlay dirs
Most of these were added back in the eos2 era[1][2] and aren't relevant
anymore. In particular, the /endless apps directory is long gone and
handling unmerged /bin, /lib and /sbin was added to support eos2.3
installs. The others (/home, /ostree, /opt and /root) simply can't be
real directories as there would be no way to maintain persistent data
there. The ostree filesystem setup has been stable for several years now
and I think we can remember to update this if we change it again.

1. https://github.com/endlessm/gnome-initial-setup/commit/a37874c79401e852b8a094d887e392f6a0e7e541
2. https://github.com/endlessm/gnome-initial-setup/commit/8c028fb114e31a9f57d3354b583d1287f0cb770b

https://phabricator.endlessm.com/T32011