overlayfs-setup: Remove special ostree/flatpak setup
This was needed when the system flatpak repo was merged with the system
ostree repo, but we're not doing that anymore. Remove that setup and add
an overlay for /sysroot/ostree. The flatpak repo overlay will be handled
by the /var overlay.

Since this new configuration won't work correctly when the repos are
still merged, check if that's the case and exit with an error if so. For
new installations this would never be the case, but a system updated
from eos3 might fail to successfully split the repos.

https://phabricator.endlessm.com/T32011