overlayfs-setup: Drop /run user_xattr mount option
On new installations or on systems that have upgraded and successfully
performed the flatpak repo split, the system flatpak repo will be in
`bare-user-only` mode. In this mode ostree doesn't use xattrs, so
there's no need to make them available for the case of allowing flatpaks
to be installed.

Since the overlayfs-setup only runs when the system flatpak repo has
been split, we can remove the call to remount `/run` with the
`user_xattr` option.

https://phabricator.endlessm.com/T32011