Remove handling of split disk systems
For EOS 4 we will no longer support split disk systems, so remove the
handling here. Primarily this involves removing the split disk mount at
`/var/endless-extra`, but it also affects a couple other tools that
needed to be aware of the mount. Some other cleanup is also removed as
it's expected that any split disk systems will never run this code.

https://phabricator.endlessm.com/T32109