Fix unprivileged distcheck
This was succeeding on Jenkins because the build runs as root and can
install and delete in `/usr/share/polkit-1/actions`. Turn the directory
into an option and override it during distcheck like the other system
directories. This also adds an error if `PKG_CHECK_VAR` fails, which
wasn't happening before.

https://phabricator.endlessm.com/T32109