eos-transient-setup: Drop support for replacing Chrome with Chromium
Since T32407, we use Chromium as the default browser in all images, so
there should be no images where booting in live mode could cause Chrome
to be downloaded unnecessarily (if it hadn’t already been downloaded by
another user).

Accordingly, we can drop the transient setup steps which prevent Chrome
from being auto-downloaded, and set Chromium as the default browser and
place it in the icon grid instead of Chrome.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T32408