tmpfiles: Don't empty directories
Per tmpfiles.d(5), the `D` directive empties the contents of the
directory when `systemd-tmpfiles` is run with `--remove`. This happens
on every boot from `systemd-tmpfiles-setup.service`, so all of these
directories were being emptied during boot. In many cases that's
probably harmless, but for `/etc/chromium-browser`, it deletes all
chromium system configuration.

https://phabricator.endlessm.com/T32863