Remove Endless Key hooks
The first version of Endless Key had:

- Windows and macOS apps to consume Kolibri & Kiwix content
- Kolibri & Kiwix content
- A bootable Endless OS image

We used magic to make the Kolibri & Kiwix content from the exFAT
partition be used by the Kolibri and Encyclopedia apps in Endless OS,
respectively.

We no longer include an Endless OS image on Endless Key; and the
encyclopedia content has been moved so even if we did still have an OS
image on the device, that part would be broken.

Remove these hooks from the OS.

(The generated run-media-eoslive.mount unit was originally added for the
benefit of these hooks, but does serve the purpose of making the host
partition on the USB device more readily accessible, so leave it in
place.)

https://phabricator.endlessm.com/T33159