Add service to set ostree bootloader on existing systems
Instead of relying on ostree's automatic bootloader detection, we'd like
to specify it since we have to carry an ugly downstream patch to make
the automatic detection do what we want. The configuration is simple
since the bootloader is either u-boot or grub. In the grub case we want
the ostree bootloader configuration to be `none` so that it only updates
the bootloader spec entries and doesn't try to manage `grub.cfg`. After
a checkpoint we can drop this script and unit.

https://phabricator.endlessm.com/T33364