Revert "psi-monitor: Drop Pressure Stall Information Monitor"
This reverts commit 356f7bb9f98feec80dbc158c36876661789ca88d.

We had replaced our in-house psi-monitor with low-memory-monitor in
enforcing mode, but it has proved too aggressive to work in production,
so we are reverting back to psi-monitor for now.

We'll try to understand why low-memory-monitor is behaving differently
than psi-monitor despite using the same PSI levels, and hopefully switch
back to low-memory-monitor again after some fine-tuning.

https://phabricator.endlessm.com/T34267