Update system CA store when /etc is updated
Users, or Endless may add some CAs which do not belong debian packages.
Those CAs will be updated into /etc/ssl/certs/ca-certificates.crt and
would be likely to be mismatched from the shipped version in /usr/etc.

Add a helper service that runs update-ca-certificates after update and
boot to sync up when it is needed.

https://phabricator.endlessm.com/T34308