eos-update-flatpak-repos: Migrate eos-runtimes to new URL
Historically the legacy runtimes were committed to the OS repo because
they were built from the same packages and therefore were entirely
deduplicated on the server. Fast forward many years and now the mixture
of ostree and flatpak commits is starting to cause problems as flatpak
is managing the repo metadata differently than flatpak.

The legacy runtimes have been duplicated into a new eos-runtimes repo
that will contain only flatpaks. In order to use it, we need to migrate
users to the new URL.

https://phabricator.endlessm.com/T33445