Add eos-esp-generator script for mounting ESP
This is intended to replace systemd-gpt-auto-generator for mounting the
ESP. The various policies we need are difficult to capture there and
would be hard to upstream. Instead, this program can capture all the
products we support to ensure the ESP is mounted where and how we need
it. It creates units in the normal generator directory, which mean they
take precedence over the gpt-auto-generator units in the late generator
directory.

I've tested this on EOS with GPT, EOS with MBR, and PAYG. What's not
currently supported is the Windows dual boot configuration where the
root disk isn't really the root disk.

https://phabricator.endlessm.com/T29930