tests: Add some useful pytest options
Personally I think the `-v` output where it shows each test on a single
line should be the default. Enabling debug logging by default is helpful
for inspecting failed tests remotely. pytest still hides the output by
default. The last one also seems like it should be the default - if you
have a test that's expected to fail but succeeds, that should be
considered a failure.

https://phabricator.endlessm.com/T29930