Add service to ensure systemd-boot installed
On systemd-boot systems, the bootloader is updated with `bootctl update`
on every boot. Unfortunately, it doesn't currently work because
`bootctl` considers systemd-boot to be installed if `/EFI/systemd` is
non-empty in the ESP. The image builder only populates `/EFI/BOOT`, so
systemd-boot is not considered installed.

If `bootctl` does not consider systemd-boot to be installed,
`eos-ensure-sd-boot` will check the `LoaderInfo` EFI variable to see if
the current bootloader is systemd-boot. If so, it runs `bootctl install`
so that the next time `bootctl update` runs, it will update the
bootloader.

https://phabricator.endlessm.com/T34703