eos-image-boot-setup: use by-uuid symlink for host device identification
It has been a rough ride to accurately translate the
endless.image.device=UUID=foo parameter into (e.g.) /dev/sda1,
because udev was presenting sda and sda3 (if unformatted) with the same
UUID.

We settled on a blkid behaviour that got us the right answer, but
it no longer works after upgrading to Debian bookworm.
blkid doesn't seem to register UUIDs for any of those devices now.

However, the udev situation has now improved, and an unformatted partition
will no longer inherit filesystem info from the parent device:
https://github.com/systemd/systemd/issues/14408

This means that we can give the by-uuid symlink approach another shot.
Both sda and sda1 will contend for this symlink, but it seems fair to
assume that the partition (which we want) will always be probed after
the parent device.

https://phabricator.endlessm.com/T35109