image-boot: use loopback mounts for all filesystems
Debian has changed to another upstream for exfatprogs which does not
have the extent-listing functionality we were relying on for image
boot from exfat.

Now that both exfat and ntfs have decent kernel drivers, we have
made the decision to avoid the complexity of the extent-mapping approach,
and simply use the kernel's loopback mount functionality within partitions
mounted with those new drivers to access the image files we are trying to
boot from.

In turn, we are able to remove /dev/mapper/endless-image-device and
the corresponding /run/media/eoslive mount which only existed because in
the extent-mapped approach it was not possible to mount the partition
holding the image file. With the loopback approach it can now be mounted
as normal.

https://phabricator.endlessm.com/T35126