image-boot: Remove device mapper, mount loopback devices directly
After removing the extent mapping approach for image booting,
we are left with a simpler code flow logic with some obvious calls
for simplification.

After we set up the loopback device, we were setting up a device-mapper
linear mapping against it. That had the function of handling non-standard
sector sizes when booting from DVD (which we no longer care about).
Then another linear mapping was placed on top when creating partitions.
Another useful function of the mapping was to give the final root
device a stable name (/dev/mapper/endless-image*), used by the
subsequent initramfs logic, and also by eos-installer when seeking
source image for installation.

Eliminating the device mapper bits for simplicity, we now set up
a symlink at /dev/disk/endless-image which points to the loop device
that we are booting from.

https://phabricator.endlessm.com/T35126