image-boot: fix read-only block device handling
The `blockdev --setro` invocation here was too broad, because in cases
of live booting with persistent storage enabled, we will need to write
to persistent.img on the host device.

You might think the next place down the stack where we could start
the read-only setting would be at the outer image mount, however,
that is again mounting the host device, and the kernel later refuses to
create a read-write mount (for persistent.img access) if the original
mount is read-only.

Leave the real partition node as read-write, and instead start the
read-only setting at the first loopback device. (All further loop devices
and mounts created within will have the read-only flag propagated.)

https://phabricator.endlessm.com/T35126