overlay-setup: Ensure units ordered after /var overlay
Trying to create a list of units that require /var as in 0cddd9e is a
losing game. It also doesn't really work properly since you can't order
against all instances of a templated unit like foo@.service.

As suggested by upstream[1], the more robust way to handle this is to
order the service before var.mount. That requires handling the /var
mount in the service, but fortunately that's straightforward for this
narrow use case.

With that change, the service will also run before
ostree-remount.service, which means it has to take care of marking the
/sysroot mount as private. Without that, the /var overlay mount will be
mirrored under /sysroot.

1. https://github.com/systemd/systemd/issues/32343

https://phabricator.endlessm.com/T35323